chore: move the engine pin past the sentinel and the whitespace ruling
The pin now reads and writes the empty-description sentinel and refuses a
marker followed by whitespace alone, so the eleven declared drift rows have
retired and are deleted here rather than left describing a window that closed.

An empty description is spellable, so an import that meets an empty `<dd>`
writes the entry and keeps its list whole. Both html-import fixtures record
that: one list, no diagnostics. `structure-split` had a single producer - a
dropped entry with a term after it - and reports nothing now.

The import ratchet moves in two directions for one reason. Two documents spell
a tree their writer output does not reproduce byte for byte, which is what
their `.fmt` fixtures record, so they leave the html round-trip set and join
the canonical fixed points at the same time: 1534 to 1536 and 663 to 661.

The definition-list test now pins what the sentinel keeps rather than what a
dropped entry broke, so it no longer cites a docs section that was removed or
a diagnostic without a producer.
