+++ resources/engine-fmt-drift.txt
@@ -14,10 +14,3 @@
 # engine-pin-drift.txt says the same thing for the same reason.
 #
 # Format: <slug><space><space><reason>
-436-an-empty-description-body-claims-no-line-below-column-0-2  the empty-body ruling (carve#1821, SS17 L3 AND FLUSH-LEFT MEANS COLUMN 0) reached the pin's READER but not its WRITER: the initial render now matches the corpus, and the writer still flattens the refused marker back to `: flush`, so re-reading its own output moves the payload into the `<dd>`. Declared here rather than in engine-pin-drift.txt because that file is about the first render, which no longer drifts.
-436-an-empty-description-body-claims-no-line-below-column-0-4  the empty-body ruling (carve#1821, SS17 L3 AND FLUSH-LEFT MEANS COLUMN 0) reached the pin's READER but not its WRITER: the initial render now matches the corpus, and the writer still flattens the refused marker back to `: flush`, so re-reading its own output moves the payload into the `<dd>`. Declared here rather than in engine-pin-drift.txt because that file is about the first render, which no longer drifts.
-438-an-empty-description-body-is-written-with-the-empty-sentinel  the empty-description sentinel (carve#1827, PART 11 SS7d): the pin READS `: {empty}` as an empty description body and reproduces the HTML, and its writer drops the description entirely instead of writing the sentinel back, so the `.fmt` fixture is the ruled output and not the pin's. Delete when the writer port ships (carve-js#1559, carve-php#1811, carve-rs#1451)
-438-an-empty-description-body-is-written-with-the-empty-sentinel-2  the empty-description sentinel (carve#1827, PART 11 SS7d): the pin READS `: {empty}` as an empty description body and reproduces the HTML, and its writer drops the description entirely instead of writing the sentinel back, so the `.fmt` fixture is the ruled output and not the pin's. Delete when the writer port ships (carve-js#1559, carve-php#1811, carve-rs#1451)
-438-an-empty-description-body-is-written-with-the-empty-sentinel-3  the empty-description sentinel (carve#1827, PART 11 SS7d): the pin READS `: {empty}` as an empty description body and reproduces the HTML, and its writer drops the description entirely instead of writing the sentinel back, so the `.fmt` fixture is the ruled output and not the pin's. Delete when the writer port ships (carve-js#1559, carve-php#1811, carve-rs#1451)
-438-an-empty-description-body-is-written-with-the-empty-sentinel-4  the empty-description sentinel (carve#1827, PART 11 SS7d): the pin READS `: {empty}` as an empty description body and reproduces the HTML, and its writer drops the description entirely instead of writing the sentinel back, so the `.fmt` fixture is the ruled output and not the pin's. Delete when the writer port ships (carve-js#1559, carve-php#1811, carve-rs#1451)
-438-an-empty-description-body-is-written-with-the-empty-sentinel-5  the empty-description sentinel (carve#1827, PART 11 SS7d): the pin READS `: {empty}` as an empty description body and reproduces the HTML, and its writer drops the description entirely instead of writing the sentinel back, so the `.fmt` fixture is the ruled output and not the pin's. Delete when the writer port ships (carve-js#1559, carve-php#1811, carve-rs#1451)
