+++ tests/html-import/empty-definition-description-not-last/expected.crv
@@ -1,6 +1,4 @@
 :: t1
-
-%%
-
+: {empty}
 :: t2
 : d2
