+++ tests/html-import/empty-definition-description/expected.crv
@@ -1,1 +1,2 @@
 :: term
+: {empty}
