+++ tests/html-import/empty-definition-description/expected.report.json
@@ -1,1 +1,1 @@
-{"mode":"safe","adapter":"generic","diagnostics":[{"code":"structure-unspellable","severity":"warning","path":"/dl[1]/dd[2]"}]}
+{"mode":"safe","adapter":"generic","diagnostics":[]}
